Login added, not finished or tested
Code should be moved and things needs to be more clean. Not tested on
real server, only dev. server. Installation install all users with the
same password, abc123, and cannot be changed right now. Some change to
the SQL structure as a password field was added to the table users.